+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## develop (unreleased)
+
+### New features
+* [#123](https://github.com/trema/pio/pull/123): Add new class `Pio::ExactMatch`.
+
+
 ## 0.11.2 (1/29/2015)
 
 ### Bugs fixed